Add data for which courses each team should have
