--- src/raw-data/test.js
@@ -1,6 +1,0 @@
-
-
-// import { hello_stuff } from '../../raw-data/person_should_have_courses';
-
-import { hello_stuff } from './person_should_have_courses';
-
